Move text utils in client code, relocate tests.
