+++ sal/plugin.py
@@ -38,7 +38,7 @@ from django.template import loader
 
 from sal.decorators import handle_access, is_global_admin
 from server.models import Machine, Plugin, MachineDetailPlugin, Report
-from server.text_utils import class_to_title
+from utils.text_utils import class_to_title
 
 
 # TODO: This can be removed along with the legacy plugin support code
